move function HasAtLeastOnePrefix into redis-shake/common
